feat: isolated the ensureSchema functions into a module and npm script
Also XOd a few files, updated README and invoked the migration runner for tests
