+++ packages/oae-activity/lib/init.js
@@ -13,8 +13,6 @@
  * permissions and limitations under the License.
  */
 
-const Cassandra = require('oae-util/lib/cassandra');
-
 const ActivityAPI = require('oae-activity');
 const ActivityPush = require('./internal/push');
 
@@ -31,37 +29,12 @@ require('./internal/notifications');
 require('./internal/email');
 
 module.exports = function(config, callback) {
-  ensureSchema(err => {
+  ActivityAPI.refreshConfiguration(config.activity, err => {
     if (err) {
       return callback(err);
     }
 
-    ActivityAPI.refreshConfiguration(config.activity, err => {
-      if (err) {
-        return callback(err);
-      }
-
-      // Configure the push notifications
-      ActivityPush.init(callback);
-    });
+    // Configure the push notifications
+    ActivityPush.init(callback);
   });
-};
-
-/**
- * Ensure that the all of the activity-related schemas are created. If they already exist, this method will not do anything.
- *
- * @param  {Function}    callback       Standard callback function
- * @param  {Object}      callback.err   An error that occurred, if any
- * @api private
- */
-const ensureSchema = function(callback) {
-  Cassandra.createColumnFamilies(
-    {
-      ActivityStreams:
-        'CREATE TABLE "ActivityStreams" ("activityStreamId" text, "activityId" text, "activity" text, PRIMARY KEY ("activityStreamId", "activityId")) WITH COMPACT STORAGE',
-      EmailBuckets:
-        'CREATE TABLE "EmailBuckets" ("bucketId" text, "userId" text, PRIMARY KEY ("bucketId", "userId"))'
-    },
-    callback
-  );
 };
