+++ packages/oae-config/lib/init.js
@@ -13,27 +13,8 @@
  * permissions and limitations under the License.
  */
 
-const Cassandra = require('oae-util/lib/cassandra');
-
 const ConfigAPI = require('oae-config');
 
 module.exports = function(config, callback) {
-  ensureSchema(() => {
-    ConfigAPI.initConfig(config, callback);
-  });
-};
-
-/**
- * Ensure that the config schema is created.
- * If both the schema and the default config exist, then this method will do nothing.
- *
- * @param  {Function}   callback    Standard callback function
- * @api private
- */
-const ensureSchema = function(callback) {
-  Cassandra.createColumnFamily(
-    'Config',
-    'CREATE TABLE "Config" ("tenantAlias" text, "configKey" text, "value" text, PRIMARY KEY ("tenantAlias", "configKey")) WITH COMPACT STORAGE',
-    callback
-  );
+  return ConfigAPI.initConfig(config, callback);
 };
