+++ packages/oae-library/lib/init.js
@@ -13,25 +13,6 @@
  * permissions and limitations under the License.
  */
 
-const Cassandra = require('oae-util/lib/cassandra');
-
 module.exports = function(config, callback) {
-  ensureSchema(callback);
-};
-
-/**
- * Ensure that the all of the library-related schemas are created. If they already exist, this method will not do anything
- *
- * @param  {Function}         callback       Standard callback function
- * @param  {Object}           callback.err   An error that occurred, if any
- * @api private
- */
-const ensureSchema = function(callback) {
-  Cassandra.createColumnFamilies(
-    {
-      LibraryIndex:
-        'CREATE TABLE "LibraryIndex" ("bucketKey" text, "rankedResourceId" text, "value" text, PRIMARY KEY ("bucketKey", "rankedResourceId")) WITH COMPACT STORAGE'
-    },
-    callback
-  );
+  return callback();
 };
