+++ packages/oae-lti/lib/migration.js
@@ -1,0 +1,20 @@
+const Cassandra = require('oae-util/lib/cassandra');
+
+/**
+ * Create the schema for Learning Tools Interoperability tools
+ *
+ * @param  {Function}   callback        Standard callback function
+ * @param  {Object}     callback.err    An error that occurred, if any
+ * @api private
+ */
+const ensureSchema = function(callback) {
+  Cassandra.createColumnFamilies(
+    {
+      LtiTools:
+        'CREATE TABLE "LtiTools" ("id" text, "groupId" text, "launchUrl" text, "secret" text, "oauthConsumerKey" text, "displayName" text, "description" text, PRIMARY KEY ("groupId", "id"))'
+    },
+    callback
+  );
+};
+
+module.exports = { ensureSchema };
