+++ packages/oae-principals/lib/init.js
@@ -17,7 +17,6 @@ const util = require('util');
 
 const AuthenticationAPI = require('oae-authentication');
 const { AuthzConstants } = require('oae-authz/lib/constants');
-const Cassandra = require('oae-util/lib/cassandra');
 const { Context } = require('oae-context');
 const TenantsAPI = require('oae-tenants');
 const { User } = require('oae-principals/lib/model');
@@ -38,53 +37,7 @@ module.exports = function(config, callback) {
   // Initialize principals delete capabilities
   require('oae-principals/lib/delete'); // eslint-disable-line import/no-unassigned-import
 
-  _ensureSchema(err => {
-    if (err) {
-      return callback(err);
-    }
-
     return _ensureGlobalAdmin(config, callback);
-  });
-};
-
-/**
- * Ensure that the all of the principal-related schemas are created. If they already exist, this method will not
- * do anything
- *
- * @param  {Function}    callback       Standard callback function
- * @param  {Object}      callback.err   An error that occurred, if any
- * @api private
- */
-const _ensureSchema = function(callback) {
-  // Both user and group information will be stored inside of the Principals CF
-  Cassandra.createColumnFamilies(
-    {
-      Principals:
-        'CREATE TABLE "Principals" ("principalId" text PRIMARY KEY, "tenantAlias" text, "displayName" text, "description" text, "email" text, "emailPreference" text, "visibility" text, "joinable" text, "lastModified" text, "locale" text, "publicAlias" text, "largePictureUri" text, "mediumPictureUri" text, "smallPictureUri" text, "admin:global" text, "admin:tenant" text, "notificationsUnread" text, "notificationsLastRead" text, "acceptedTC" text, "createdBy" text, "created" timestamp, "deleted" timestamp)',
-
-      // Map an email address to user ids. An e-mail address can be used by *multiple* users
-      PrincipalsByEmail:
-        'CREATE TABLE "PrincipalsByEmail" ("email" text, "principalId" text, PRIMARY KEY ("email", "principalId"))',
-
-      // Map a user id to a desired email address and verification token
-      PrincipalsEmailToken:
-        'CREATE TABLE "PrincipalsEmailToken" ("principalId" text PRIMARY KEY, "email" text, "token" text)',
-
-      // Track user visits to groups they are members of
-      UsersGroupVisits:
-        'CREATE TABLE "UsersGroupVisits" ("userId" text, "groupId" text, "latestVisit" text, PRIMARY KEY ("userId", "groupId"))'
-    },
-    err => {
-      if (err) {
-        return callback(err);
-      }
-      Cassandra.runQuery(
-        'CREATE INDEX IF NOT EXISTS ON "Principals" ("tenantAlias")',
-        [],
-        callback
-      );
-    }
-  );
 };
 
 /**
