+++ packages/oae-tenants/lib/init.js
@@ -13,37 +13,9 @@
  * permissions and limitations under the License.
  */
 
-const Cassandra = require('oae-util/lib/cassandra');
 const TenantsAPI = require('oae-tenants');
 
 module.exports = function(config, callback) {
-  // Create the schema and register
-  ensureSchema(err => {
-    if (err) {
-      return callback(err);
-    }
-
-    // Initialize the middleware that will add the tenant onto the request
-    TenantsAPI.init(config.servers, callback);
-  });
-};
-
-/**
- * Ensure that the tenant schema is created. If the tenant schema has not been created, or the default tenant has not been seeded,
- * both will be performed automatically. If both the schema and the default tenant exist, then this method will do nothing.
- *
- * @param  {Function}  callback     Standard callback function
- * @api private
- */
-const ensureSchema = function(callback) {
-  Cassandra.createColumnFamilies(
-    {
-      Tenant:
-        'CREATE TABLE "Tenant" ("alias" text PRIMARY KEY, "displayName" text, "host" text, "emailDomains" text, "countryCode" text, "active" boolean)',
-      TenantNetwork: 'CREATE TABLE "TenantNetwork" ("id" text PRIMARY KEY, "displayName" text);',
-      TenantNetworkTenants:
-        'CREATE TABLE "TenantNetworkTenants" ("tenantNetworkId" text, "tenantAlias" text, "value" text, PRIMARY KEY ("tenantNetworkId", "tenantAlias")) WITH COMPACT STORAGE;'
-    },
-    callback
-  );
+  // Initialize the middleware that will add the tenant onto the request
+  TenantsAPI.init(config.servers, callback);
 };
